Improve the break appeal calculation code.
The break_appeal_ member in NGBoxFragmentBuilder was a bit fuzzy, did
too many things, and didn't behave entirely correctly.

Don't let it deal with the break appeal of early-breaks anymore; instead
move this into the NGEarlyBreak object (which also allows us to remove
this member from the rare data object in NGLayoutResult).

Also start off with perfect break appeal (rather than assuming a
last-resort break) in the builder, and downgrade it if we have to,
rather than upgrading it to "perfect" manually in FinishFragmentation().
We'd fail to realize that break appeal was perfect if we had a
break-before break token for a float inside, for instance.

Correctly propagate the break appeal all the way up to the
fragmentainer.

Apart from fixing a few "corner cases", this is needed by nested block
fragmentation (in a follow-up CL), so that we can push the first
fragment of an inner multicol to the next outer fragmentainer, to avoid
violations (e.g. breaking inside something with break-inside:avoid).

Bug: 829028
Change-Id: I4ec2444f623efddf399a34e604a3b1375a171438